chore: move title.tsx
